organize common components
