+++ src/common/design/colors.ts
@@ -12,13 +12,11 @@ export const RED = '#FB2F3C';
 export const YELLOW_200 = '#FFC453';
 export const BLUE_200 = '#4E81E3';
 export const GREEN_200 = '#3CB78B';
-export const PURPLE_100 = '#C3B2FF';
-export const PURPLE_300 = '#5335B4';
 export const PINK = '#FFD7D5';
 export const CREAM = '#F9F1E1';
-export const BRAND_PRIMARY = '#A578FF';
-export const BRAND_SECONDARY = GRAY_600;
-export const BRAND_INFO = '#4199E1';
+export const PRIMARY = '#A578FF';
+export const ACCENT_1 = GRAY_600;
+export const ACCENT_2 = GRAY_700;
 export const BRAND_SUCCESS = YELLOW_200;
 export const BRAND_WARNING = YELLOW_200;
 export const BRAND_DANGER = RED;
