+++ src/common/design/index.ts
@@ -1,2 +1,5 @@
+export * from './icons';
 export * from './styles';
 export * from './themes';
+export * from './breakpoints';
+export * from './colors';
