+++ src/common/design/styles/background-colors.scss
@@ -1,3 +1,5 @@
+@import './colors.scss';
+
 .bg-gray-100 {
   background-color: $gray-100;
 }
