+++ src/common/design/styles/index.ts
@@ -1,1 +1,1 @@
-export * from './base-styles';
+export * from './base-styles.scss';
