+++ src/common/design/styles/typography/text-colors.scss
@@ -1,3 +1,5 @@
+@import '../colors.scss';
+
 .color-gray-900 {
   color: $gray-900;
 }
