+++ src/common/utils/dates/date-display-type.ts
@@ -1,0 +1,10 @@
+export type DateDisplayType =
+  | 'year'
+  | 'month'
+  | 'month-short'
+  | 'date'
+  | 'day'
+  | 'day-short'
+  | 'full'
+  | 'month dd, yyyy'
+  | 'mm/dd/yyyy';
