+++ src/common/utils/dates/get-random-date-in-range.ts
@@ -1,0 +1,7 @@
+export const getRandomDateInRange = (
+  startDate: Date,
+  endDate: Date
+): number => {
+  const _diff: number = endDate.getTime() - startDate.getTime();
+  return new Date(Math.random() * _diff + startDate.getTime()).getTime();
+};
