+++ src/common/utils/dates/get-random-date.ts
@@ -1,0 +1,3 @@
+export const getRandomDate = (): number => {
+  return new Date(Math.random() * new Date().getTime()).getTime();
+};
