+++ src/common/utils/index.ts
@@ -1,4 +1,4 @@
+export * from './dates';
+export * from './display';
 export * from './regex';
 export * from './type-guards';
-export * from './display/commafy-number';
-export * from './display/get-dollar-string';
